cdp/e2e-test/pyproject.toml
[project]
name = "cdp-e2e-test"
version = "0.1.0"
description = "CDP E2E Test Framework with Playwright"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytest",
    "playwright",
    "allure-pytest",
    "pytest-playwright",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "pytest-cov",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests", "src/e2e/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--alluredir=reports/allure-results",
]
markers = [
    "slow: marks tests as slow",
    "smoke: marks tests as smoke tests",
    "login: marks tests that require login",
]

[tool.allure]
link-name = "CDP E2E Test"

[tool.hatch.build.targets.wheel]
packages = ["src/e2e"]

[dependency-groups]
dev = [
    "openpyxl>=3.1.5",
]
